Python files after ruff linting and formatting
The ruff formatter has now been run on the entire
project. The linter was run afterwards and no
errors were raised.

The file make_demo_hdf5.py has been added to the resources
folder. The file allows users to make a demo data set to try
with hdf5view and follow the tutorials in the documentation
(coming later) with the same data.