Added documentation
The documentation is written in reStructuredText within the
Sphinx framework and uses the pydata sphonx theme.

Removed old "doc" folder - all the images it contained
are now in docs/_static_readme, and the README.md has
been updated so that github and pypi should be able to
find the images for the readme.

Other unneeded images have been deleted.

The templates in the _templates folder are based on those from
kikuchipy.

.pre-commit-config.yaml has been updated to ignore docs/conf.py,
which was causing an error because it doesn't contain a docstring.