feat(TreeView and TableView behaviour)
The Tree View has been modified so that the width of the columns
always fits to the contents. This makes it easier when expanding a
tree to see the names of the datasets or groups.

The Table View has been modified so that for single column datasets,
the width of the column always fits the contents. A dataset containing
a single string will then be displayed so that the whole string can be
read, for example. For datasets with more than one column, the column
width is set to a standard size for a nicer appearance and the user
can adjust the size if desired.